- dll은 지원하지 않을거라.. 이제 TGON_API 안쓸거 - UIText 구현... 텍스트 정렬까지 구현해놓음 - 기타 등등

+++ Source/TGON/Core/RTTI.h
@@ -126,7 +126,7 @@ public:
 
 inline size_t hash<tgon::RTTI>::operator()(const tgon::RTTI& rhs) const noexcept
 {
-    return static_cast<size_t>(rhs.GetHashCode());
+    return rhs.GetHashCode();
 }
     
 } /* namespace std */
